Move deprecation mode from xpack core to deprecation module.
